GPB-2 Refactor data management into own module
