+++ src/main/resources/application.conf
@@ -7,6 +7,7 @@ ktor {
             de.genpare.modules.SetupKt.setup
             de.genpare.modules.HelloWorldKt.helloWorld
             de.genpare.modules.MemberManagementKt.memberManagement
+            de.genpare.modules.DataManagementKt.dataManagement
         ]
     }
 }
